+++ src/constants.ts
@@ -2,4 +2,4 @@
  * ANY_RESOURCE is used to describe any Google Cloud resource.
  * This includes all assets both ingested and not ingested with this integration
  */
-export const ANY_RESOURCE = 'ANY_RESOURCE';
+export const ANY_RESOURCE = 'resource';