+++ src/steps/cloud-asset/constants.ts
@@ -2,6 +2,7 @@ import {
   generateRelationshipType,
   RelationshipClass,
 } from '@jupiterone/integration-sdk-core';
+import { ANY_RESOURCE } from '../../constants';
 import { MULTIPLE_J1_TYPES_FOR_RESOURCE_KIND } from '../../utils/iamBindings/resourceKindToTypeMap';
 import { J1_TYPE_TO_KEY_GENERATOR_MAP } from '../../utils/iamBindings/typeToKeyGeneratorMap';
 import {
@@ -79,14 +80,25 @@ const bindingResourceTargetTypes = Object.keys(J1_TYPE_TO_KEY_GENERATOR_MAP)
  * IAM policies can target any resource in Google Cloud. Because we do not ingest every resource,
  * we have chosen, instead, to represent the relationship as IAM Binding assigned to ANY_RESOURCE.
  */
-export const BINDING_ALLOWS_ANY_RESOURCE_RELATIONSHIPS =
-  bindingResourceTargetTypes.map((resourceType) => {
+export const BINDING_ALLOWS_ANY_RESOURCE_TYPE = generateRelationshipType(
+  RelationshipClass.ALLOWS,
+  bindingEntities.BINDINGS._type,
+  ANY_RESOURCE,
+);
+export const BINDING_ALLOWS_ANY_RESOURCE_RELATIONSHIPS = [
+  {
+    // Needed for documentation
+    _class: RelationshipClass.HAS,
+    _type: BINDING_ALLOWS_ANY_RESOURCE_TYPE,
+    sourceType: API_SERVICE_ENTITY_TYPE,
+    targetType: ANY_RESOURCE,
+    indexMetadata: {
+      enabled: false,
+    },
+  },
+  ...bindingResourceTargetTypes.map((resourceType) => {
     return {
-      _type: generateRelationshipType(
-        RelationshipClass.ALLOWS,
-        bindingEntities.BINDINGS._type,
-        resourceType,
-      ),
+      _type: BINDING_ALLOWS_ANY_RESOURCE_TYPE,
       sourceType: bindingEntities.BINDINGS._type,
       _class: RelationshipClass.ALLOWS,
       targetType: resourceType,
@@ -94,21 +106,34 @@ export const BINDING_ALLOWS_ANY_RESOURCE_RELATIONSHIPS =
         enabled: false,
       },
     };
-  });
+  }),
+];
 
-export const API_SERVICE_HAS_ANY_RESOURCE_RELATIONSHIPS =
-  bindingResourceTargetTypes.map((resourceType) => {
+export const API_SERVICE_HAS_ANY_RESOURCE_TYPE = generateRelationshipType(
+  RelationshipClass.HAS,
+  API_SERVICE_ENTITY_TYPE,
+  ANY_RESOURCE,
+);
+export const API_SERVICE_HAS_ANY_RESOURCE_RELATIONSHIPS = [
+  {
+    // Needed for documentation
+    _class: RelationshipClass.HAS,
+    _type: API_SERVICE_HAS_ANY_RESOURCE_TYPE,
+    sourceType: API_SERVICE_ENTITY_TYPE,
+    targetType: ANY_RESOURCE,
+    indexMetadata: {
+      enabled: false,
+    },
+  },
+  ...bindingResourceTargetTypes.map((resourceType) => {
     return {
       _class: RelationshipClass.HAS,
-      _type: generateRelationshipType(
-        RelationshipClass.HAS,
-        API_SERVICE_ENTITY_TYPE,
-        resourceType,
-      ),
+      _type: API_SERVICE_HAS_ANY_RESOURCE_TYPE,
       sourceType: API_SERVICE_ENTITY_TYPE,
       targetType: resourceType,
       indexMetadata: {
         enabled: false,
       },
     };
-  });
+  }),
+];