+++ src/steps/cloud-asset/index.test.ts
@@ -19,6 +19,8 @@ import {
   fetchResourceManagerFolders,
   fetchResourceManagerOrganization,
   fetchResourceManagerProject,
+  FOLDER_ENTITY_TYPE,
+  ORGANIZATION_ENTITY_TYPE,
   PROJECT_ENTITY_TYPE,
 } from '../resource-manager';
 import {
@@ -217,6 +219,7 @@ function createMockContext() {
     // Temporary tweak to make this test pass since its recording has been updated from the new organization/v3
     instanceConfig: {
       ...integrationConfig,
+      projectId: 'j1-gc-integration-dev-v3',
       serviceAccountKeyFile: integrationConfig.serviceAccountKeyFile.replace(
         'j1-gc-integration-dev-v2',
         'j1-gc-integration-dev-v3',
@@ -254,6 +257,10 @@ describe('#fetchIamBindings', () => {
       await fetchResourceManagerFolders(context);
       await fetchResourceManagerProject(context);
       await buildOrgFolderProjectMappedRelationships(context);
+
+      // Used for making direct relationships for google_cloud_api_service_has_resource
+      await fetchStorageBuckets(context);
+
       await fetchIamCustomRoles(context);
       await fetchIamManagedRoles(context);
       await fetchIamServiceAccounts(context);
@@ -292,25 +299,8 @@ describe('#fetchIamBindings', () => {
         everyone_assigned_google_iam_role,
         google_iam_role_assigned_role,
 
-        google_cloud_api_service_has_storage_bucket,
-        google_cloud_api_service_has_bigquery_dataset,
-        google_cloud_api_service_has_kms_crypto_key,
-        google_cloud_api_service_has_kms_key_ring,
-        google_cloud_api_service_has_function,
-        google_cloud_api_service_has_project,
-        google_cloud_api_service_has_folder,
-        google_cloud_api_service_has_organization,
-        google_cloud_api_service_has_service,
-
-        google_iam_binding_allows_cloud_organization,
-        google_iam_binding_allows_cloud_folder,
-        google_iam_binding_allows_cloud_project,
-        google_iam_binding_allows_storage_bucket,
-        google_iam_binding_allows_bigquery_dataset,
-        google_iam_binding_allows_kms_crypto_key,
-        google_iam_binding_allows_kms_key_ring,
-        google_iam_binding_allows_billing_account,
-        google_iam_binding_allows_cloud_function,
+        google_cloud_api_service_has_resource,
+        google_iam_binding_allows_resource,
       } = separateGraphObjectsByType(
         context.jobState.collectedRelationships,
         context.jobState.encounteredTypes,
@@ -324,36 +314,30 @@ describe('#fetchIamBindings', () => {
         'google_iam_binding_assigned_service_account',
       );
       expect(
-        google_iam_binding_allows_cloud_project,
+        google_iam_binding_allows_resource,
       ).toHaveBothDirectAndMappedRelationships(
-        'google_iam_binding_allows_cloud_project',
+        'google_iam_binding_allows_resource',
+      );
+      expect(
+        google_cloud_api_service_has_resource,
+      ).toHaveBothDirectAndMappedRelationships(
+        'google_cloud_api_service_has_resource',
       );
       expect(
         google_iam_binding_uses_role,
       ).toHaveBothDirectAndMappedRelationships('google_iam_binding_uses_role');
 
-      // Currently, do not have examples of some resources ingested in this integration and some in others
-      // For now, just check that we have some relationships
-      expect(google_iam_binding_allows_storage_bucket.length > 0).toBe(true);
-      expect(google_iam_binding_allows_bigquery_dataset.length > 0).toBe(true);
-      expect(google_iam_binding_allows_kms_crypto_key.length > 0).toBe(true);
-      expect(google_iam_binding_allows_kms_key_ring.length > 0).toBe(true);
-      expect(google_iam_binding_allows_billing_account.length > 0).toBe(true);
-      expect(google_iam_binding_allows_cloud_function.length > 0).toBe(true);
-
-      expect(google_cloud_api_service_has_storage_bucket.length > 0).toBe(true);
-      expect(google_cloud_api_service_has_bigquery_dataset.length > 0).toBe(
-        true,
-      );
-      expect(google_cloud_api_service_has_kms_crypto_key.length > 0).toBe(true);
-      expect(google_cloud_api_service_has_kms_key_ring.length > 0).toBe(true);
-      expect(google_cloud_api_service_has_function.length > 0).toBe(true);
-
       // Ensure google_cloud_api_service HAS any resource in the organization resource hierarchy relationships are NOT created
-      expect(google_cloud_api_service_has_project).toBeUndefined();
-      expect(google_cloud_api_service_has_folder).toBeUndefined();
-      expect(google_cloud_api_service_has_organization).toBeUndefined();
-      expect(google_cloud_api_service_has_service).toBeUndefined();
+      expect(
+        google_cloud_api_service_has_resource.filter((r: any) => {
+          r._toEntityKey?.startsWith('organizations/');
+        }).length,
+      ).toBe(0);
+      expect(
+        google_cloud_api_service_has_resource.filter((r: any) => {
+          r._toEntityKey?.startsWith('folders/');
+        }).length,
+      ).toBe(0);
 
       // Mapped Relationships
       expect(google_iam_binding_assigned_user).toHaveOnlyMappedRelationships(
@@ -380,15 +364,30 @@ describe('#fetchIamBindings', () => {
       expect(google_iam_role_assigned_role).toBeUndefined();
 
       // Direct Relationships
+      // Organization relationships
       expect(
-        google_iam_binding_allows_cloud_organization,
-      ).toHaveOnlyDirectRelationships(
-        'google_iam_binding_allows_cloud_organization',
-      );
+        google_iam_binding_allows_resource.filter((r: any) =>
+          r._toEntityKey?.startsWith('organizations/'),
+        ).length,
+      ).toBeGreaterThan(0),
+        expect(
+          google_iam_binding_allows_resource.filter(
+            (r: any) =>
+              r._mapping?.targetEntity?._type === ORGANIZATION_ENTITY_TYPE,
+          ).length,
+        ).toBe(0),
+        // Folder relationships
+        expect(
+          google_iam_binding_allows_resource.filter((r: any) =>
+            r._toEntityKey?.startsWith('folders/'),
+          ).length,
+        ).toBeGreaterThan(0);
       expect(
-        google_iam_binding_allows_cloud_folder,
-      ).toHaveOnlyDirectRelationships('google_iam_binding_allows_cloud_folder');
-      // These are for "Convienence Values" which we make sure connected Basic Role exists in `createBasicRolesForBindings` so it will always be direct.
+        google_iam_binding_allows_resource.filter(
+          (r: any) => r._mapping?.targetEntity?._type === FOLDER_ENTITY_TYPE,
+        ).length,
+      ).toBe(0);
+      // These are for "Convienence Members"
       expect(google_iam_binding_assigned_role).toHaveOnlyDirectRelationships(
         'google_iam_binding_assigned_role',
       );
@@ -567,22 +566,22 @@ describe('#fetchIamBindings', () => {
 
         expect(
           bindingAnyResourceMappedRelationships.filter(
-            (r) => r._type === 'google_iam_binding_allows_storage_bucket',
+            (r) => r._mapping.targetEntity._type === 'google_storage_bucket',
           ).length > 0,
         ).toBe(true);
         expect(
           bindingAnyResourceMappedRelationships.filter(
-            (r) => r._type === 'google_iam_binding_allows_bigquery_dataset',
+            (r) => r._mapping.targetEntity._type === 'google_bigquery_dataset',
           ).length > 0,
         ).toBe(true);
         expect(
           bindingAnyResourceMappedRelationships.filter(
-            (r) => r._type === 'google_iam_binding_allows_cloud_function',
+            (r) => r._mapping.targetEntity._type === 'google_cloud_function',
           ).length > 0,
         ).toBe(true);
         expect(
           bindingAnyResourceMappedRelationships.filter(
-            (r) => r._type === 'google_iam_binding_allows_cloud_project',
+            (r) => r._mapping.targetEntity._type === 'google_cloud_project',
           ).length > 0,
         ).toBe(true);
         expect(bindingAnyResourceMappedRelationships).toTargetEntities(
@@ -603,6 +602,11 @@ describe('#createBasicRolesForBindings', () => {
   it('should truncate permission values for both bindings and roles', async () => {
     await withRecording('createBasicRolesForBindings', __dirname, async () => {
       const context = createMockContext();
+
+      // Needed for gathering folder and organization level IAM bindings
+      await fetchResourceManagerOrganization(context);
+      await fetchResourceManagerFolders(context);
+
       await fetchIamManagedRoles(context);
       await fetchIamBindings(context);
       await createBasicRolesForBindings(context);